handle login with errors and jwt support
